Added a support for valid rank configuration.
In the bSocks config, there is now a list of valid ranks and a default rank option.
If a player is not in a valid group their group is set to the default group.

Removed static methods and added access through the bSocks singleton instance.
Removed FigAdmin support as we dont use that anymore, code is commented out for future reference.